pyproject.toml
[tool.poetry]
name = "haruka-bot-re"
description = "Push dynamics and live informations from bilibili to QQ. Based on nonebot2."
version = "0.1.0"
authors = ["GrandDuke1106 <gao166901@outlook.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/GrandDuke1106/HarukaBot-Re"
repository = "https://github.com/GrandDuke1106/HarukaBot-Re/tree/master/src/plugins/haruka_bot_re"
documentation = "https://github.com/GrandDuke1106/HarukaBot-Re#readme"
keywords = ["nonebot", "nonebot2", "qqbot", "bilibili", "bot"]


[tool.poetry.dependencies]
python = ">=3.8,<4.0"
click = ">=8.1.3"
httpx = ">=0.24.1"
nonebot-adapter-onebot = ">=2.2.3"
nonebot-plugin-apscheduler = ">=0.3.0"
nonebot2 = {extras = ["fastapi"], version = ">=2.0.0"}
playwright = ">=1.35.0"
pydantic = ">=1.10.9, <=2.0"
python-dotenv = ">=1.0.0"
tortoise-orm = {extras = ["asyncpg"], version = ">=0.19.3"}
bilireq = ">=0.2.6"
packaging = ">=23.1"
nonebot-plugin-guild-patch = ">=0.2.3"
msvc-runtime = ">=14.34.31931" # Conditional dependencies are not supported by Poetry, so handle it in other ways
#aunly-captcha-solver = ">=0.1"
dynamicadaptor = ">=0.4.7"
dynrender-skia = ">=0.1.6"

[tool.poetry.scripts]
hb = "haruka_bot_re.__main__:main"

[tool.poetry.dev-dependencies]
nonebot-plugin-gocqhttp = ">=0.6.10"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["haruka_bot_re", "nonebot_plugin_gocqhttp", "nonebot_plugin_guild_patch"]
plugin_dirs = []
builtin_plugins = []

[tool.pdm.version]
source = "file"
path = "haruka_bot_re/version.py"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"